feat: moved intervention review to lower level
