+++ src/app/pages/costEffectiveness/interventionReview/interventionReview.component.scss
